moved method addOrderToCustomerByTheirId from OrderController to CustomerController, fixed context
