+++ src/main/java/waterDelivery/domain/Customer.java
@@ -40,4 +40,7 @@ public class Customer {
     public Customer() {
     }
 
+    public void addOrder(Order order) {
+        orderList.add(order);
+    }
 }
